+++ admin/products/delete_image.php
@@ -1,10 +1,18 @@
 <?php
-include '../db.php';
-include '../users/auth.php';
+include_once('../db.php');
+include_once( '../users/auth.php');
 
-$id = intval($_GET['id']);
-$conn->query("DELETE FROM product_images WHERE key_image = $id");
+if ($_SESSION["role"] != "admin") {
+	echo "<script>alert('You do not have access to delete a record');history.back();</script>";
+	exit;
+}
 
-// Return updated image list
-include 'get_images.php';
+$key_proudct = $_GET['key_product'];
+if (isset($_GET['id'])) {
+	$id = intval($_GET['id']);
+	$conn->query("DELETE FROM product_images WHERE key_image = $id");
+	// Return updated image list
+	include("get_images.php");
+}
+
 ?>
