+++ admin/youtube_gallery/delete.php
@@ -1,5 +1,18 @@
-<?php include '../db.php';
-$id = $_GET['id'];
-$conn->query("DELETE FROM youtube_gallery WHERE key_youtube_gallery=$id");
+<?php 
+include '../db.php';
+include '../users/auth.php';
+
+if ($_SESSION["role"] != "admin") {
+	echo "<script>alert('You do not have access to delete a record');history.back();</script>";
+	exit;
+}
+
+
+if (isset($_GET['id'])) {
+	$id = $_GET['id'];
+	$conn->query("DELETE FROM youtube_gallery WHERE key_youtube_gallery=$id");
+}
+
 header("Location: list.php");
+exit;
 ?>
